Rationalize internal structure of map page a bit
