+++ src/components/page-info.ts
