+++ src/pages/map/index.page.ts
@@ -1,7 +1,7 @@
 import h from "@macrostrat/hyper";
 import { ClientOnly } from "~/renderer/client-only";
 
-const MapPage = () => import("./map");
+const MapPage = () => import("./map-interface");
 
 export function Page() {
   return h(
